move game_phase to timewiz
