+++ utils/timewiz/game_phase.py
@@ -1,0 +1,11 @@
+from datetime import datetime, timezone
+
+from . import CW_OFFSET, CW_PERIODS, SPEED
+
+
+def game_phase():
+    adjustment = -37.0
+    utcdt = datetime.now(timezone.utc)
+    cwtdt = datetime.fromtimestamp(SPEED * (utcdt.timestamp() + CW_OFFSET), tz=timezone.utc)
+    cwadt = datetime.fromtimestamp(cwtdt.timestamp() + SPEED * adjustment, tz=timezone.utc)
+    return f'{CW_PERIODS[cwadt.hour//6]}'
